making links from chapter 1 open in a new tab
These edits are only for chapter 1 and add `target='_blank'` to <a> tags that don't have them.

I'm happy to go through other chapters to do the same if it is helpful. I would also be happy to discuss an implementation that affects all links to avoid manual edits in the future.